+++ lib/src/test/java/io/ably/lib/test/realtime/RealtimeConnectFailTest.java
@@ -16,8 +16,8 @@ import io.ably.lib.rest.Auth.TokenCallback;
 import io.ably.lib.rest.Auth.TokenDetails;
 import io.ably.lib.rest.Auth.TokenParams;
 import io.ably.lib.test.common.Setup;
+import io.ably.lib.test.common.Helpers.ConnectionWaiter;
 import io.ably.lib.test.common.Setup.TestVars;
-import io.ably.lib.test.realtime.Helpers.ConnectionWaiter;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
 import io.ably.lib.types.ErrorInfo;
