+++ lib/src/test/java/io/ably/lib/test/realtime/RealtimeConnectTest.java
@@ -10,9 +10,9 @@ import org.junit.BeforeClass;
 import io.ably.lib.realtime.AblyRealtime;
 import io.ably.lib.realtime.ConnectionState;
 import io.ably.lib.test.common.Setup;
+import io.ably.lib.test.common.Helpers.CompletionWaiter;
+import io.ably.lib.test.common.Helpers.ConnectionWaiter;
 import io.ably.lib.test.common.Setup.TestVars;
-import io.ably.lib.test.realtime.Helpers.CompletionWaiter;
-import io.ably.lib.test.realtime.Helpers.ConnectionWaiter;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
 import io.ably.lib.util.Log;
