+++ lib/src/test/java/io/ably/lib/test/realtime/RealtimeMessageTest.java
@@ -12,11 +12,11 @@ import io.ably.lib.realtime.AblyRealtime;
 import io.ably.lib.realtime.Channel;
 import io.ably.lib.realtime.ChannelState;
 import io.ably.lib.test.common.Setup;
+import io.ably.lib.test.common.Helpers.ChannelWaiter;
+import io.ably.lib.test.common.Helpers.CompletionSet;
+import io.ably.lib.test.common.Helpers.CompletionWaiter;
+import io.ably.lib.test.common.Helpers.MessageWaiter;
 import io.ably.lib.test.common.Setup.TestVars;
-import io.ably.lib.test.realtime.Helpers.ChannelWaiter;
-import io.ably.lib.test.realtime.Helpers.CompletionSet;
-import io.ably.lib.test.realtime.Helpers.CompletionWaiter;
-import io.ably.lib.test.realtime.Helpers.MessageWaiter;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
 import io.ably.lib.types.ErrorInfo;
