+++ lib/src/test/java/io/ably/lib/test/realtime/RealtimePresenceHistoryTest.java
@@ -15,12 +15,12 @@ import io.ably.lib.rest.AblyRest;
 import io.ably.lib.rest.Auth;
 import io.ably.lib.rest.Auth.TokenParams;
 import io.ably.lib.test.common.Setup;
+import io.ably.lib.test.common.Helpers.ChannelWaiter;
+import io.ably.lib.test.common.Helpers.CompletionSet;
+import io.ably.lib.test.common.Helpers.CompletionWaiter;
+import io.ably.lib.test.common.Helpers.PresenceWaiter;
+import io.ably.lib.test.common.Helpers.RawProtocolWaiter;
 import io.ably.lib.test.common.Setup.TestVars;
-import io.ably.lib.test.realtime.Helpers.ChannelWaiter;
-import io.ably.lib.test.realtime.Helpers.CompletionSet;
-import io.ably.lib.test.realtime.Helpers.CompletionWaiter;
-import io.ably.lib.test.realtime.Helpers.PresenceWaiter;
-import io.ably.lib.test.realtime.Helpers.RawProtocolWaiter;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
 import io.ably.lib.types.PaginatedResult;
