+++ lib/src/test/java/io/ably/lib/test/realtime/RealtimePresenceTest.java
@@ -13,12 +13,12 @@ import io.ably.lib.rest.AblyRest;
 import io.ably.lib.rest.Auth;
 import io.ably.lib.rest.Auth.TokenParams;
 import io.ably.lib.test.common.Setup;
+import io.ably.lib.test.common.Helpers.ChannelWaiter;
+import io.ably.lib.test.common.Helpers.CompletionSet;
+import io.ably.lib.test.common.Helpers.CompletionWaiter;
+import io.ably.lib.test.common.Helpers.ConnectionWaiter;
+import io.ably.lib.test.common.Helpers.PresenceWaiter;
 import io.ably.lib.test.common.Setup.TestVars;
-import io.ably.lib.test.realtime.Helpers.ChannelWaiter;
-import io.ably.lib.test.realtime.Helpers.CompletionSet;
-import io.ably.lib.test.realtime.Helpers.CompletionWaiter;
-import io.ably.lib.test.realtime.Helpers.ConnectionWaiter;
-import io.ably.lib.test.realtime.Helpers.PresenceWaiter;
 import io.ably.lib.types.AblyException;
 import io.ably.lib.types.ClientOptions;
 import io.ably.lib.types.PaginatedResult;
